feat: create useStatusList hook and code reorg
